fap-discord-bot/requirements.txt
discord.py>=2.3.2
patchright>=1.40.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
aiofiles>=23.0.0
python-dotenv>=1.0.0
nodriver>=1.0.0
